backend/requirements.txt
fastapi>=0.116.1
pydantic>=2.11.7
uvicorn>=0.24.0
python-dotenv>=1.0.0
pillow>=10.0.0
aiofiles>=23.2.1
google-generativeai>=0.1.0
python-multipart>=0.0.20